split ssr to ssr and ssr-replacement
effector-react/ssr contains some methods to implement forking on server
effector-react/ssr-replacement can be used to replace effector-react on server
