+++ packages/effector-react/ssr-replacement.js
@@ -1,0 +1,3 @@
+//@flow
+
+export * from '../../src/react/ssr-replacement'
